Extract uthash wrappers into header
